float

+++ C++/Examples/ControlModule/c_control_lqrArthur_vel.c
@@ -189,9 +189,9 @@ pv_type_actuation c_control_lqrArthur_vel_controller(pv_msg_input inputData){
 
 	printf("control output numCols = %d\n",control_output.numCols);
 	printf("control output numRows = %d\n",control_output.numRows);
-	int (*ptr)[4];
+	float (*ptr)[4];
 	  ptr = control_output.pData;
-	printf("%d %d %d\n", **ptr, *(*(ptr + 0) + 1), *(*(ptr + 0) + 2));
+	printf("%f %f %f\n", **ptr, *(*(ptr + 0) + 1), *(*(ptr + 0) + 2));
 	//
 	//The result must be in a struct pv_msg_io_actuation
 	actuation_signals.escRightNewtons= (float)control_output.pData[0];
